requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-model-utils==4.3.1
environs==9.5.0
idna==3.4
marshmallow==3.20.1
packaging==23.2
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
telebot==0.0.5
tzdata==2023.3
urllib3==2.0.6
